feat: add ESLint configuration and Prettier formatting
- Added ESLint configuration with modern JavaScript rules
- Added Prettier configuration for consistent code formatting
- Updated package.json with new dependencies and scripts
- Refactored code to follow ESLint standards
- Updated README with new package name and usage instructions
- Added bun.lock for package management

BREAKING CHANGE: Package name changed from 'rmdir-cli' to '@involvex/rmdir-cli

+++ bin/rmdir-cli.js
@@ -1,9 +1,18 @@
 #!/usr/bin/env node
 
-'use strict';
-var rmdir = require('../index');
+"use strict";
+var rmdir = require("../index");
+var version = require("../package.json").version;
 
-
-if(process.argv[2]){
+if (process.argv[2] == "--help" || process.argv[2] == "-h") {
+  console.log("Usage: rmdir <dir>");
+  console.log("rmdir <dir> - Recursively removes a directory.");
+  process.exit(0);
+} else if (process.argv[2] == "--version" || process.argv[2] == "-v") {
+  console.log(version);
+  process.exit(0);
+} else {
+  if (process.argv[2]) {
     rmdir(process.argv[2]);
+  }
 }
